creates wodDetailDialog
