+++ src/components/pages/mainPage.tsx
@@ -1,11 +1,7 @@
 import React, { useState } from 'react';
-import {
-    Autocomplete,
-    Dialog, DialogContent, DialogTitle, Grid, styled, TextField
-} from '@mui/material';
+import { Grid, styled } from '@mui/material';
 import WodCardList from '../lists/wodCardList';
-import { movementOptionsData } from '../../helpers/hardcodedData';
-import { MovementOptions } from '../../interfaces/dataInterfaces';
+import WodDetailDialog from '../dialogs/wodDetailDialog';
 
 const classesPrefix = 'mainPage';
 
@@ -27,11 +23,6 @@ const StyledGrid = styled(Grid)(() => {
 
 const MainPage = (): JSX.Element => {
     const [isDialogOpen, setIsDialogOpen] = useState(false);
-    const [selectedMovementOne, setSelectedMovementOne] = useState<MovementOptions | null>(null);
-
-    const handleMovementChange = (selectedOption: MovementOptions | null): void => {
-        setSelectedMovementOne(selectedOption);
-    };
 
     const handleCardClick = () => {
         setIsDialogOpen(true);
@@ -49,46 +40,10 @@ const MainPage = (): JSX.Element => {
             {
                 isDialogOpen
                 && (
-                    <Dialog
-                        open={isDialogOpen}
-                        onClose={() => { handleDialogClose(); }}
-                        fullWidth
-                    >
-                        <DialogTitle>Workout Details</DialogTitle>
-                        <DialogContent>
-                            <Grid container>
-                                <Grid item xs={6}>
-                                    <Autocomplete
-                                        size='small'
-                                        disabled={false}
-                                        options={movementOptionsData}
-                                        getOptionLabel={(option): string => {
-                                            return option.label;
-                                        }}
-                                        onChange={(event, newValue): void => {
-                                            handleMovementChange(newValue);
-                                        }}
-                                        value={selectedMovementOne}
-                                        renderInput={(params): JSX.Element => {
-                                            return (
-                                                <TextField
-                                                    {...params}
-                                                    label='Movement One'
-                                                    fullWidth
-                                                    margin='normal'
-                                                    variant='filled'
-                                                    inputProps={{
-                                                        ...params.inputProps,
-                                                        autoComplete: 'off'
-                                                    }}
-                                                />
-                                            );
-                                        }}
-                                    />
-                                </Grid>
-                            </Grid>
-                        </DialogContent>
-                    </Dialog>
+                    <WodDetailDialog
+                        isDialogOpen={isDialogOpen}
+                        handleDialogClose={handleDialogClose}
+                    />
                 )
             }
         </StyledGrid>
